introduce new test_layers.py module
