+++ dask/tests/test_distributed.py
@@ -6,15 +6,15 @@ distributed = pytest.importorskip("distributed")
 import asyncio
 from functools import partial
 from operator import add
-import sys
 
 from tornado import gen
 
 import dask
 from dask import persist, delayed, compute
+import dask.bag as db
 from dask.delayed import Delayed
 from dask.utils import tmpdir, get_named_args
-from distributed import futures_of, SchedulerPlugin, Client
+from distributed import futures_of
 from distributed.client import wait
 from distributed.utils_test import (  # noqa F401
     gen_cluster,
@@ -24,7 +24,6 @@ from distributed.utils_test import (  # noqa F401
     loop,
     client as c,
     varying,
-    cleanup,
 )
 
 
@@ -118,7 +117,6 @@ def test_fused_blockwise_dataframe_merge(c, fuse):
 
 
 def test_futures_to_delayed_bag(c):
-    db = pytest.importorskip("dask.bag")
     L = [1, 2, 3]
 
     futures = c.scatter([L, L])
@@ -289,6 +287,71 @@ async def test_annotations_blockwise_unpack(c, s, a, b):
     assert_eq(z, np.ones(10) * 4.0)
 
 
+@pytest.mark.parametrize(
+    "io",
+    [
+        "ones",
+        "zeros",
+        "full",
+    ],
+)
+@pytest.mark.parametrize("fuse", [True, False])
+def test_blockwise_array_creation(c, io, fuse):
+    np = pytest.importorskip("numpy")
+    da = pytest.importorskip("dask.array")
+
+    chunks = (5, 2)
+    shape = (10, 4)
+
+    if io == "ones":
+        darr = da.ones(shape, chunks=chunks)
+        narr = np.ones(shape)
+    elif io == "zeros":
+        darr = da.zeros(shape, chunks=chunks)
+        narr = np.zeros(shape)
+    elif io == "full":
+        darr = da.full(shape, 10, chunks=chunks)
+        narr = np.full(shape, 10)
+
+    darr += 2
+    narr += 2
+    with dask.config.set({"optimization.fuse.active": fuse}):
+        darr.compute()
+        da.assert_eq(darr, narr)
+
+
+@gen_cluster(client=True)
+async def test_blockwise_numpy_args(c, s, a, b):
+    """Test pack/unpack of blockwise that includes a NumPy literal argument"""
+    da = pytest.importorskip("dask.array")
+    np = pytest.importorskip("numpy")
+
+    def fn(x, dt):
+        assert type(dt) is np.uint16
+        return x.astype(dt)
+
+    arr = da.blockwise(
+        fn, "x", da.ones(1000), "x", np.uint16(42), None, dtype=np.uint16
+    )
+    res = await c.compute(arr.sum(), optimize_graph=False)
+    assert res == 1000
+
+
+@gen_cluster(client=True)
+async def test_blockwise_numpy_kwargs(c, s, a, b):
+    """Test pack/unpack of blockwise that includes a NumPy literal keyword argument"""
+    da = pytest.importorskip("dask.array")
+    np = pytest.importorskip("numpy")
+
+    def fn(x, dt=None):
+        assert type(dt) is np.uint16
+        return x.astype(dt)
+
+    arr = da.blockwise(fn, "x", da.ones(1000), "x", dtype=np.uint16, dt=np.uint16(42))
+    res = await c.compute(arr.sum(), optimize_graph=False)
+    assert res == 1000
+
+
 @gen_cluster(client=True)
 async def test_combo_of_layer_types(c, s, a, b):
     """Check pack/unpack of a HLG that has every type of Layers!"""
@@ -335,93 +398,3 @@ async def test_annotation_pack_unpack(c, s, a, b):
     unpacked_hlg = HighLevelGraph.__dask_distributed_unpack__(packed_hlg, annotations)
     annotations = unpacked_hlg["annotations"]
     assert annotations == {"workers": {"n": ("alice",)}}
-
-
-class SchedulerImportCheck(SchedulerPlugin):
-    """Plugin to help record which modules are imported on the scheduler"""
-
-    def __init__(self, pattern):
-        self.pattern = pattern
-
-    async def start(self, scheduler):
-        # Record the modules that have been imported when the scheduler starts
-        self.start_modules = set()
-        for mod in set(sys.modules):
-            if not mod.startswith(self.pattern):
-                self.start_modules.add(mod)
-            else:
-                # Maually remove the target library
-                sys.modules.pop(mod)
-
-
-def get_start_modules(dask_scheduler):
-    import_check_plugins = [
-        p for p in dask_scheduler.plugins if type(p) is SchedulerImportCheck
-    ]
-    assert len(import_check_plugins) == 1
-
-    plugin = import_check_plugins[0]
-    return plugin.start_modules
-
-
-def _dataframe_shuffle():
-    pd = pytest.importorskip("pandas")
-    dd = pytest.importorskip("dask.dataframe")
-
-    # Perform a computation using an HLG-based shuffle
-    df = pd.DataFrame({"a": range(10), "b": range(10, 20)})
-    return dd.from_pandas(df, npartitions=2)
-
-
-def _dataframe_broadcast_join():
-    pd = pytest.importorskip("pandas")
-    dd = pytest.importorskip("dask.dataframe")
-
-    # Perform a computation using an HLG-based broadcast join
-    df = pd.DataFrame({"a": range(10), "b": range(10, 20)})
-    ddf1 = dd.from_pandas(df, npartitions=4)
-    ddf2 = dd.from_pandas(df, npartitions=1)
-    return ddf1.merge(ddf2, how="left", broadcast=True)
-
-
-def _array_creation():
-    da = pytest.importorskip("dask.array")
-
-    # Perform a computation using HLG-based array creation
-    return da.ones((100,)) + da.zeros((100,))
-
-
-@pytest.mark.parametrize(
-    "op,lib",
-    [
-        (_dataframe_shuffle, "pandas."),
-        (_dataframe_broadcast_join, "pandas."),
-        (_array_creation, "numpy."),
-    ],
-)
-@pytest.mark.parametrize("optimize_graph", [True, False])
-def test_scheduler_highlevel_graph_unpack_import(op, lib, optimize_graph, loop):
-    # Test to highlight that in some cases we import modules like pandas
-    # on the scheduler when unpacking some HighLevelGraphs.
-    # This is a problem because numpy/pandas may not be installed in the
-    # process where the scheduler is running -- leading to an ImportError.
-
-    with cluster(scheduler_kwargs={"plugins": [SchedulerImportCheck(lib)]}) as (
-        scheduler,
-        workers,
-    ):
-        with Client(scheduler["address"], loop=loop) as c:
-            # Perform a computation using a HighLevelGraph Layer
-            c.compute(op(), optimize_graph=optimize_graph)
-
-            # Get the new modules which were imported on the scheduler during the computation
-            end_modules = c.run_on_scheduler(lambda: set(sys.modules))
-            start_modules = c.run_on_scheduler(get_start_modules)
-            new_modules = end_modules - start_modules
-
-            # Check that the scheduler didn't start with `lib`
-            # (otherwise we arent testing anything)
-            assert not any(module.startswith(lib) for module in start_modules)
-
-            # Check whether we imported `lib` on the scheduler
-            assert not any(module.startswith(lib) for module in new_modules)
